Combine actions and util - player, added YoutubeApi for commentThreads
